+++ src/main/java/com/api/soccerTournament/controller/PlayerController.java
@@ -15,19 +15,26 @@ public class PlayerController {
     private PlayerService playerService;
 
     @GetMapping(path = "/getAll", produces = Const.responseFormat)
-    public Response readAll() {
+    public Response<Person> readAll() {
         Response response = playerService.readAll();
         return response;
     }
 
     @PostMapping(path = "/write", produces = Const.responseFormat)
-    public Response write(@RequestBody Person person){
-        Response response = playerService.write(person);
+    public Response write(@RequestBody Person player){
+        if (player == null) return new Response(Const.statusCodeFailParamNull, "player");
+        if (player.name == null) return new Response(Const.statusCodeFailParamNull, "player.name");
+        if (player.name.trim().length() == 0) return new Response(Const.statusCodeFailParamEmpty, "player.coach");
+
+        Response response = playerService.write(player);
         return response;
     }
 
     @PostMapping(path = "/delete", produces = Const.responseFormat)
     public Response delete(@RequestBody Integer id) {
+        if (id == null) return new Response(Const.statusCodeFailParamNull, "id");
+        if (id <= 0) return new Response(Const.statusCodeFailParamInvalid, "id");
+
         Response response = playerService.delete(id);
         return response;
     }
